% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amgut1.filt}
\alias{amgut1.filt}
\title{American Gut Project Data (Filtered)}
\format{
A data frame with samples as rows and microbial taxa as columns.
}
\source{
SpiecEasi package
}
\usage{
data(amgut1.filt)
}
\description{
This dataset, `amgut1.filt`, is a filtered dataset from the American Gut Project.
It contains microbiome data suitable for network analysis.
}
\examples{
data(amgut1.filt)
head(amgut1.filt)
}
\keyword{datasets}
