% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cclasso.R
\name{cclasso}
\alias{cclasso}
\title{Correlation Inference for Compositional Data through Lasso}
\usage{
cclasso(x, counts, pseudo, k_cv, lam_int, k_max, n_boot)
}
\arguments{
\item{x}{A numeric matrix of dimension n x p, where n is the number of samples and p is the number of compositional variables.}

\item{counts}{Logical, indicating whether the compositional data matrix is a count matrix. Default is FALSE.}

\item{pseudo}{Numeric, the pseudo count added if `counts = TRUE`. Default is 0.5.}

\item{k_cv}{Integer, the number of folds for cross-validation. Default is 3.}

\item{lam_int}{Numeric vector of length 2, specifying the interval for the tuning parameter lambda. Default is `c(1e-4, 1)`.}

\item{k_max}{Integer, the maximum number of iterations for the golden section method. Default is 20.}

\item{n_boot}{Integer, the number of bootstrap replications. Default is 20.}
}
\value{
A list containing:
\item{var_w}{Variance estimation.}
\item{cor_w}{Correlation estimation.}
\item{p_vals}{P-values for elements of `cor_w` indicating if they are equal to 0.}
\item{lambda}{The final tuning parameter.}
\item{info_cv}{Information for cross-validation.}
}
\description{
The main function to infer correlations for compositional data using Lasso.
}
