% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMiNet.R
\name{CMiNet}
\alias{CMiNet}
\title{This function constructs a consensus network from microbiome data using multiple methods.}
\usage{
CMiNet(
  data,
  quantitative = TRUE,
  TT,
  pearson = list(enabled, params = list()),
  spearman = list(enabled, params = list()),
  bicor = list(enabled, params = list()),
  sparcc = list(enabled, params),
  spiecEasi_mb = list(enabled, params),
  spiecEasi_glasso = list(enabled, params),
  spring = list(enabled, params),
  gcoda = list(enabled, params),
  c_MI = list(enabled, params),
  cclasso = list(enabled, params)
)
}
\arguments{
\item{data}{A numeric matrix or data frame of microbiome data where rows represent samples and columns represent features.}

\item{quantitative}{A logical value indicating if the data is quantitative. Defaults to TRUE.}

\item{TT}{A numeric value representing the quantile threshold used to binarize the adjacency matrices.}

\item{pearson}{A list with parameters for Pearson correlation (`enabled`, `params`).}

\item{spearman}{A list with parameters for Spearman correlation (`enabled`, `params`).}

\item{bicor}{A list with parameters for biweight midcorrelation (`enabled`, `params`).}

\item{sparcc}{A list with parameters for SparCC correlation (`enabled`, `params`).}

\item{spiecEasi_mb}{A list with parameters for SpiecEasi Meinshausen-Bühlmann method (enabled, params).}

\item{spiecEasi_glasso}{A list with parameters for SpiecEasi Graphical Lasso method (enabled, params).}

\item{spring}{A list with parameters for SPRING method (`enabled`, `params`).}

\item{gcoda}{A list with parameters for GCODA method (`enabled`, `params`).}

\item{c_MI}{A list with parameters for Conditional Mutual Information method (`enabled`, `params`).}

\item{cclasso}{A list with parameters for CCLasso method (`enabled`, `params`).}
}
\value{
A list containing the consensus network, edge list, and any errors encountered during network construction.
}
\description{
This function constructs a consensus network from microbiome data using multiple methods.
}
