% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_MI.R
\name{conditional_MI}
\alias{conditional_MI}
\title{Conditional Mutual Information Network Construction}
\usage{
conditional_MI(data, q1, q2, quantitative)
}
\arguments{
\item{data}{A numeric matrix where rows represent genes and columns represent samples. If \code{quantitative} is TRUE, data will be log-transformed.}

\item{q1}{A numeric value representing the quantile threshold for filtering edges in the order 0 adjacency matrix.}

\item{q2}{A numeric value representing the quantile threshold for filtering edges in the order 1 adjacency matrix.}

\item{quantitative}{A logical value indicating if the data is quantitative. If TRUE, data is log-transformed.}
}
\value{
A list containing:
  \item{G_order0}{The adjacency matrix after order 0 calculation.}
  \item{G_order1}{The adjacency matrix after order 1 calculation.}
  \item{Gval_order0}{The CMI values for each gene pair in the order 0 adjacency matrix.}
  \item{Gval_order1}{The CMI values for each gene pair in the order 1 adjacency matrix.}
  \item{quantile_order0}{The quantile threshold used for filtering the order 0 adjacency matrix.}
  \item{quantile_order1}{The quantile threshold used for filtering the order 1 adjacency matrix.}
  \item{sum_order0}{The sum of edges in the order 0 adjacency matrix.}
  \item{sum_order1}{The sum of edges in the order 1 adjacency matrix.}
}
\description{
Constructs a network using conditional mutual information (CMI) to identify relationships between genes.
The function applies two levels of CMI computation, using quantile thresholds to filter edges in the resulting adjacency matrices.
}
\details{
The function computes the conditional mutual information for gene pairs, applying two levels of calculation:
  \enumerate{
    \item \strong{Order 0}: Computes CMI between each pair of genes and filters edges based on \code{q1}.
    \item \strong{Order 1}: Refines the network by conditioning on shared neighbors, filtering edges based on \code{q2}.
  }
}
