% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vis_FinalNet.R
\name{plot_network}
\alias{plot_network}
\title{Plot Network in High Resolution and Interactive Format}
\usage{
plot_network(
  network_final,
  node_color = "skyblue",
  edge_color = "grey",
  label_color = "black"
)
}
\arguments{
\item{network_final}{A binary adjacency matrix where values greater than a specified threshold are set to 1, indicating the presence of an edge.}

\item{node_color}{Character. Color of the nodes in the plot. Default is \code{"skyblue"}.}

\item{edge_color}{Character. Color of the edges in the plot. Default is \code{"grey"}.}

\item{label_color}{Character. Color of the node labels in the plot. Default is \code{"black"}.}
}
\value{
A \code{visNetwork} interactive plot of the network is displayed in the R console.
}
\description{
This function generates a network plot from a binary adjacency matrix using the \code{igraph} and \code{visNetwork} packages.
It saves a high-resolution static image of the network and provides an interactive visualization.
}
\details{
The function creates a static plot saved as a high-resolution PNG file and also generates an interactive network plot using \code{visNetwork}.
Nodes and edges are colored based on the specified parameters, and labels are positioned within the nodes.
The interactive plot highlights nodes and edges upon selection.
}
