% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparCC.count.R
\name{SparCC.count}
\alias{SparCC.count}
\title{SparCC for Counts Data}
\usage{
SparCC.count(x, imax, kmax, alpha, Vmin)
}
\arguments{
\item{x}{A numeric matrix where rows represent samples and columns represent variables (count data).}

\item{imax}{Integer, the number of resampling times from the posterior distribution. Default is 20.}

\item{kmax}{Integer, the maximum iteration steps for SparCC. Default is 10.}

\item{alpha}{Numeric, the threshold for strong correlation. Default is 0.1.}

\item{Vmin}{Numeric, the minimal variance if negative variance appears. Default is 1e-4.}
}
\value{
A list containing:
\item{cov.w}{The estimated covariance matrix.}
\item{cor.w}{The estimated correlation matrix.}
}
\description{
This function performs SparCC analysis for count data to estimate covariance and correlation matrices.
}
